Renamed iomediator.py to windowgrabber.py, because, after all cleanups, it only contains the WindowGrabber class.
